requirements.txt
fastapi[all]
numpy
pillow
tensorflow